ccsp_tr069_pa_mapper_arm.xml: convert dos -> unix
Mixing Unix and DOS line endings in the same file is a problem as
some editors may try to automatically "fix" the inconsistency. That
leads to spurious file changes, problems applying patches, etc.

ccsp_tr069_pa_mapper_arm.xml is DOS format, but a few Unix lines were
added recently as part of:

  11b3a7f TCXB6-9071 : TR069 query fails for Secondary Radius Server parameters

An earlier version of this patch converted the whole file to dos
format. This is version 2, which converts the file to unix instead.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>