fix crash CcspCwmppoMpaGetParameterNames() when name is invalid ( fix for Alias Manager )
CcspCwmppoMpaGetParameterNames() calls CcspTr069FreeStringQueue() for
ParamList unconditionally on function exit, therefore ParamList must
be initialised before any chance to branch to the function exit via
a call to "goto EXIT2" etc.

Move initialisation of ParamList to the start of the function.

For completeness move initialisation of StringAllocList to the start
of the function too, and apply the same change to other similar
functions.

Verify that:

  GPN of empty string ("") with "NextLevel": true, returns "Device" :false
  GPN of "Device.ManagementServer.PeriodicInformInterval" throws error with faultcode of 9003 and faultString as "Invalid arguments"

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>