OFW-21 : Device.ManagementServer.URL should have value of configured URL if ACSOverride=0 in boot config file
Bug Description:
The value(URL) stored in the parameter fails to load after DUT is rebooted. It takes up the default URL that is there is the bootconfig file.

Fix includes:
Code changes done to ensure that the URL set gets reflected when DUT is rebooted and connection to that ACS URL is also established.

signed off by: Subhradeep Nath <subhradeep.nath@libertyglobal.com>

fix comparisons between a pointer and the character value 0

Fix compiler warnings.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>