minor code cleanup in CcspCwmppoGetAcsInfo()
Use sizeof(pProperty->AcsUrl) as the size limit for the new URL
instead of assuming the buffer size equals CCSP_CWMP_MAX_URL_SIZE (it
does so now but it's more robust to use sizeof() than to hardcode).

If the new URL is too long then ignore it. The previous approach was
to truncate but that just results in an invalid URL.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>