return an error for tr069 GPV from the internal data model ( OFW-375 )
Fix for OFW-375 (GPV expected to fail with WiFi indexes 1 and 2 for TR-069).

If paramter passed to GPV is an internal instance of Data Model (ie
not a CWMP instance), return error instead of returning SUCCESS with
empty string. This check is already in place for SPV.

Tests done:
Verified that 'Invalid parameter name' is returned for GPV of parameters with internal index.
 "Device.WiFi.Radio.1.ChannelsInUse".
 "Device.WiFi.Radio.2.ChannelsInUse"
 "Device.WiFi.SSID.1.BSSID"
 "Device.WiFi.SSID.16.BSSID"
 "Device.WiFi.AccessPoint.1.Enable"

Verified that below  parameters with CWMP Index return right results for GPV from ACS
 "Device.WiFi.Radio.10000.ChannelsInUse": "1,2,3,4,5,6,7,8,9,10,11,12,13"
 "Device.WiFi.Radio.10100.ChannelsInUse": "36,40,44,48,52,56,60,64,100,104,108,112,116,120,124,128"
 "Device.WiFi.SSID.10001.BSSID": "38:43:7D:8A:47:75"
 "Device.WiFi.SSID.10008.BSSID": "3A:43:7D:8A:47:75"
 "Device.WiFi.AccessPoint.10001.Enable": 1
  Device.WiFi.AccessPoint.10008.Enable: 0

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>