fix for segfault due to concurrent execution ( OFW-742 )
There is a possible segfault if a request times out.

Current architecture does not terminate the request handler thread
on timeouts and it leads to crashes.

Setting the timeout value to infinite. For better responsiveness,
the timeouts can be handled in ccsp-common-library.

Also avoid unnecessary repeated zeroing of intHttpContent in
CcspCwmpAcscoRequest(). The struct was being zero'ed 3 times before
use.