Download RPC: Download RPC implementation ( OFW-2210 ) - Part 1
Fix Description:
Implement Download RPC logic in CcspTR069Pa by calling the FirmwareDownload
parameters with values which receives from ACS RPC request

Download RPC request fields,
1. Username: empty. Assuming HTTPS or controlled HTTP use only, this is out-of-scope
2. Password: empty. Assuming HTTPS or controlled HTTP use only, this is out-of-scope
3. FileSize: set to 0, out of scope
4. TargetFilename: empty. Out of scope, this is meant to instruct under what name the file
should be stored on the local system. Assumption is that the file is locally stored under
the name as it is downloaded
5. DelaySeconds: set to 0, non-zero value out of scope. This implies that the download
must be started immediately. If it is not possible to start the download within an hour after
receiving the request, the CPE must send a TransferComplete reply indicating failure with FaultCode 9010 (see TransferComplete)
6. SuccesURL: not used, field to be empty
7. FailureURL: not used, field to be empty

Validation Performed:
1. Trigger RPC with following inputs
{'Username': '', 'DelaySeconds': 5, 'TargetFileName': '', 'FailureURL': '', 'FileSize': 200, 'SuccessURL': '',
'URL': 'http://10.15.215.5/F5685_2.11.13-2112.9_2022-05-19T15-39-21.pkgtb', 'CommandKey': '', 'Password': '', 'FileType': '1 Firmware Upgrade Image'}

2. ACS should get DownloadResponse immediately sent after Download request has been
received and the process is executing. Downloadresponse must be sent in the same CWMP session
as the Download RPC Request

Upstream Candidate: Maybe. LGI Download RPC is very limited and most of the fields are not implemented.

Signed-off-by: Sarveswaran Nagarajan <sanagarajan.contractor@libertyglobal.com>

fix memory leak in CcspCwmpsoMcoDownload_PrepareArgs()

pSlapVar is allocated twice and used for both pParamValueArray[0] and
pParamValueArray[1]

Upstream Candidate: No. This is on top of LGI commit
This can be squashed with the following commit
https://bitbucket.upc.biz/projects/RMC/repos/ccsp-tr069-pa/commits/3861ef6031e6aafab3dac7f2c7aa2516b3feae2e

Signed-off-by: Cibil <cpankiras.contractor@libertyglobal.com>