fix No fault code returned when RebootRPC/ScheduleInform is executed with invalid argument length ( OFW-3966 )
BUG DESCRIPTION:
For RebootRPC/ScheduleInform the argument size limit as per requirement is 32 characters.

FIX DESCRIPTION:
The maximum length of the argument to be taken from ACS was set to 64,
Therefore have reduced the max length to 32 as per MVXREQ-194.

TESTING PROCEDURE:
1.Provision the Gateway
2.Make the DUT online in ACS.
3.Execute Reboot RPC from ACS by providing the CommandKey as "InvalidCommandKeybghjkstwueytrsbb"

Upstream Candidate: Yes, RebootRPC/ScheduleInform argument size is as
perTR-069 standard.