Revert "fix modem unreachable via TR-069 ( OFW-2495 F3896LG-760 )" ( OFW-3614 )
fix DUT is offline in ACS when the ConnectionRequest port is changed immediately after factory reset

When port change is happening immediately after factory reset, the connection request session is not completed,
before that the value change signal thread is started and the value change event is added to event array along with the active session events.
and processing of the value change event is not happened.

rmcs:
The original fix was specific to TLN enviornment where the eRouter Ipv6
address used to get changed during lease expiries.
Reverting the commit done for OFW-2495 with the confirmation that TLN Ipv6 network is stable
without the fix.

This reverts commit b28b2a5f8f7bc2d5dece8768053107f7423de790.

Signed-off-by: Sreeja S Kumar <srekumar.contractor@libertyglobal.com>