RDKBDEV-2032 : REVERSE_INULL
Reason for change:Null-checking "pMyObject->AcsUrl" suggests that it may be null, but it has already been dereferenced on all paths leading to the check.
Null-checking "pWmpSession" suggests that it may be null, but it has already been dereferenced on all paths leading to the check.
Test Procedure: Sanity Testing
Risks: None
Signed-off-by: sugumar@infosys.com
Change-Id: Ifac399a23d6808c72729b828a276af11b7020f22