RDKBDEV-2055 : Deadcode and reverse null
Reason for change:In line no 1988 & 1864 Execution cannot reach this statement: "bFilteredOut = 0;".
Null-checking "pCwmpFault" suggests that it may be null, but it has already been dereferenced on all paths leading to the check.
Null-checking "val" suggests that it may be null, but it has already been dereferenced on all paths leading to the check.
Test Procedure: Sanity Testing
Risks: None
Signed-off-by: sugumar@infosys.com
Change-Id: Id413b2ed85fcf0eeb4f43f299e22602c5103c8fe