TransferComplete packets are not seen in the DownloadRPC-Retry scenario ( OFW-4497 F5685SI-1298 )
[Issue]

TransferComplete packets are not seen for download rpc retry scenario after 1hr of retry.

[Description]

Transfercomplete packets are not seen for retry scenario due to if conditional check failure. Retry is happening for every minute till reach out the timeout value of 60mins and after that the iRetrycount was resetted to "0" because of that if conditional check fails since it is checking for iRetrycount isequalto "60".

[Solution]

Done changes in the conditional check to trigger the transfer complete after 1hr of retry for download rpc retry scenario.

[Testing]

1. Ensure DUT is online in ACS
2. Trigger Download RPC when image located server is down
3. CPE should retry to download firmware download
4. Wait for an hour
4. Take the packet capture
5. Check the TransferComplete message and got correct code:<cwmp:TransferComplete>
                <CommandKey/>
                <FaultStruct>
                    <FaultCode>
                        9010
                        </FaultCode>
                    <FaultString>
                        Download failure
                        </FaultString>
                    </FaultStruct>

Upstream Candidate: Maybe

Signed-off-by: Saravanakumar Karthikeyan <skarthikeyan.contractor@libertyglobal.com>