Device.RootDataModelVersion is returning a value 1 instead of the actual TR181 DM version (OFW-4027)
[Description]
Device.RootDataModelVersion is returning a value 1 instead of the TR181 DataModel version the
OFW DataModel is based on. Currently the TR181 standard data model on which the TR181 is based on
is 2.13

Sagemcom also added support for fetching the Device.RootDataModelVersion through dmcli. This
patch is not required. This datamodel is only for data representation on the ACS. The patch on MV3
deviates from all the other OFW gateways. Keeping them aligned.

[Solution]
Change the DM version to 2.13. Drop the patch for MV3.

[Testing Performed]
1. Boot up the gateway
2. Once the device is online on the ACS, query the Device.RootDataModelVersion parameter.
3. Ensure that the parameter returns the DM version as 2.13.

Upstream Candidate: No. LGI specific changes.

Signed-off-by: Ajay Vishnu Bhat <abhat@libertyglobal.com>