TransferComplete packets are missing for second retry scenario after changing the https ACS url to http ( OFW-5037 )
[Issue]
TransferComplete packets are not sent to ACS, when there is a change in ACS url.

[Description]
When there is a change in ACS url then, the firmware download status value change parameter entry was removed from PSM db (eRT.com.cisco.spvtg.ccsp.tr069pa.PmNotif.Device.DeviceInfo.X_RDKCENTRAL-COM_FirmwareDownloadStatus). Due to this, the value change was not monitored by tr069 and the transfer complete was not sent to ACS.

[Solution]
As a fix, notification of "Device.DeviceInfo.X_RDKCENTRAL-COM_FirmwareDownloadStatus" was explicitly set to "1" on ACS URL change.

[Testing]
1.Ensure DUT is online in ACS
2.Trigger Download RPC when image located server is down
3.CPE should retry to download the firmware
4.wait for 1hr and the transfer complete was sent to ACS with fault code after 1hr of retry.
5.Change the ACS URL to dmcli eRT setv Device.ManagementServer.URL string http://ACSDevCloud.lab.nl.dmdsdp.com:9675/dawn/CPEManager/CPEs/genericTR69
6.Follow steps 2 to 4 again
7.Check the TransferComplete message and got correct code:cwmp:TransferComplete
<CommandKey/>
<FaultStruct>
<FaultCode>
9010
</FaultCode>
<FaultString>
Download failure
</FaultString>
</FaultStruct>

Signed-off-by: Saravanakumar Karthikeyan <skarthikeyan.contractor@libertyglobal.com>