DownloadRPC scenarios are failing after trying invalid URL case ( OFW-5337 F5685SI-1782 )
[Issue]

DownloadRPC is failing with 9001 Request Denied after invalid URL scenario.

[Description]

Triggered downloadRPC call from ACS with URL as http:// and this will fail with 9001 Request Denied.Then tried downloadRPC call with correct URL AND image name, but it also shows error with 9001.
This is because download in progress is not reset to false during first scenerio and stays in download pending for previous operation.

[Solution]

As a fix made download in progress is reset to false.

[Testing]

1. Upgrade the image with new change
2. Try downloadRPC from ACS with URL as http://  --> this will fail with 9001 Request Denied
3. Try downloadRPC with correct URL AND image name --> successful and new image flashed

Signed-off-by: Sarveswaran Nagarajan <sanagarajan.contractor@libertyglobal.com>