gix downloadRPC scenarios are failing after fwupgrademanager component restart ( OFW-4891 F5685SI-1468 )
[Issue]

After fwupgrademanager component restarted, status datamodel loses its notification attribute which leads
to not sending out any TC after component restart

[Description]

Tried downloadRPC in this order- with corrupted/half copied image, restricted, unavailable, available
images after fwupgrademanager restarted but everything is failing from ACS, since status parameter value is not updating.
It is because after fwupgrademanager restarted status parameter attribute value goes to zero and
it is not passing value change in parameter. Because of this Transfer Complete event is also not sending here.

[Solution]

Done changes to fetch status parameter attribute value and set the notification if in case it is disabled.

[Testing]

1. Upgrade the image with new change
2. Kill fwupgrademanager and wait for it to come up.
3. Execute downloadrpc in this order - with corrupted/half copied image, restricted, unavailable, available images ---------> Get expected results

Signed-off-by: Anagha prakash <apadincharachalukulangara.ext@libertyglobal.com>