add additional log message when current Firmware version is same as the requested firmware image ( F5685SI-2209 OFW-5924 )
[Issue]
When current Firmware version is same as the requested firmware image, two log messages are included instead of one

[Description]
Incorrect log message is displayed when current Firmware version is same as the requested firmware image
"Firmware download is failed with status Request Denied" message is also seen apart from "Current FW is same, Ignoring request".

[Solution]
Modifed the logic to ignore the addition log message.

[Testing]
Upgrade the image, and confirm that correct log msg is diplayed under Network Log on GUI,
when current firmwarename is same as the requested one.

Signed-off-by: Sarveswaran Nagarajan <sanagarajan.contractor@libertyglobal.com>