fix log message when current Firmware version is same as the requested firmware image ( OFW-5924 )
[Issue]
When current Firmware version is same as the requested firmware
it displying the log message as error.

[Description]
Incorrect log message is displayed when current Firmware version
is same as the requested firmware image.
It should display the log message priority as notice.

[Solution]
Changed the priority of log message from error to notice.

[Testing]
Upgrade the image, and confirm that correct log msg is diplayed
under Network Log on GUI,
when current firmwarename is same as the requested one.

Signed-off-by: Vishal Chandrakanth <vchandrakanth.ext@libertyglobal.com>